spearmint: Fix wave random at high level-times

+++ code/renderercommon/tr_common.h
@@ -150,7 +150,7 @@ extern	cvar_t	*r_fontForceAutoHint;
 qboolean	R_GetModeInfo( int *width, int *height, float *windowAspect, int mode );
 
 float R_NoiseGet4f( float x, float y, float z, double t );
-int R_RandomOn( float t );
+int R_RandomOn( double t );
 void  R_NoiseInit( void );
 
 void	R_LoadImage( const char *name, int *numLevels, textureLevel_t **pic );
